requirements.txt
aiogram>=3.0
asyncpg
sqlalchemy
aiofiles
aiohttp
pydantic-settings
python-dotenv
pytest
pytest-asyncio
httpx